Refactor NetworksDropdown, split it into components
