+++ packages/nextjs/components/NetworksDropdown/NetworksDropdown.tsx
@@ -1,11 +1,5 @@
 import { useEffect, useRef, useState } from "react";
-import Image from "next/image";
-import * as wagmiChains from "@wagmi/core/chains";
-import { useTheme } from "next-themes";
-import Select, { MultiValue, OptionProps, SingleValue, components } from "react-select";
-import { Chain } from "viem";
-import { EyeIcon, PlusIcon, WrenchScrewdriverIcon, XMarkIcon } from "@heroicons/react/24/outline";
-import { useGlobalState } from "~~/services/store/store";
+import { CustomOption, getIconComponent } from "./CustomOption";
 import {
   GroupedOptions,
   Options,
@@ -13,27 +7,19 @@ import {
   filterChains,
   formDataToChain,
   getStoredChainsFromLocalStorage,
-  isChainStored,
   mapChainsToOptions,
   removeChainFromLocalStorage,
   storeChainInLocalStorage,
-} from "~~/utils/abi-ninja/networksDropdownUtils";
+} from "./utils";
+import * as wagmiChains from "@wagmi/core/chains";
+import { useTheme } from "next-themes";
+import Select, { MultiValue, SingleValue } from "react-select";
+import { Chain } from "viem";
+import { XMarkIcon } from "@heroicons/react/24/outline";
+import { useGlobalState } from "~~/services/store/store";
 import { getPopularTargetNetworks, notification } from "~~/utils/scaffold-eth";
 
 type Chains = Record<string, Chain>;
-
-const getIconComponent = (iconName: string | undefined) => {
-  switch (iconName) {
-    case "EyeIcon":
-      return <EyeIcon className="h-6 w-6 mr-2 text-gray-500" />;
-    case "localhost":
-      return <WrenchScrewdriverIcon className="h-6 w-6 mr-2 text-gray-500" />;
-    case "PlusIcon":
-      return <PlusIcon className="h-6 w-6 mr-2 text-gray-500" />;
-    default:
-      return <Image src={iconName || "/mainnet.svg"} alt="default icon" width={24} height={24} className="mr-2" />;
-  }
-};
 
 const networks = getPopularTargetNetworks();
 const initialGroupedOptions = networks.reduce<GroupedOptions>(
@@ -97,39 +83,6 @@ const filteredChains = Object.keys(unfilteredChains)
   }, {} as Chains);
 
 const networkIds = new Set(networks.map(network => network.id));
-
-const { Option } = components;
-
-type CustomOptionProps = OptionProps<Options, false, { label: string; options: Options[] }> & {
-  onDelete: (chain: Options) => void;
-};
-const CustomOption = (props: CustomOptionProps) => {
-  const { data } = props;
-  const handleDelete = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    props.onDelete(data);
-  };
-
-  return (
-    <Option {...props}>
-      <div className="flex items-center justify-between">
-        <div className="flex items-center">
-          {typeof data.icon === "string" ? getIconComponent(data.icon) : data.icon}
-          {data.label}
-        </div>
-
-        {isChainStored(data) && (
-          <div
-            className="h-4 w-4 text-red-500 cursor-pointer font-bold flex items-center justify-center"
-            onClick={handleDelete}
-          >
-            ✕
-          </div>
-        )}
-      </div>
-    </Option>
-  );
-};
 
 export const NetworksDropdown = ({ onChange }: { onChange: (options: any) => any }) => {
   const [isMobile, setIsMobile] = useState(false);
